+++ src/p2p/src/my_node/src/node_list.py
@@ -6,7 +6,7 @@
 # Imports
 from random import choice
 
-from . import LocalNode, PeerNode
+from .. import LocalNode, PeerNode
 
 
 # Definitions
